Fix longtable madness
It's wild. Apparently, longtable still refuses to interact well with other floats on the page---I've found bug reports on this from 2003!has not managed to inter

And for some crazy reason, the Pandoc developers think it is a good idea to hardcode longtable usage into the table rendering.

Total madness!